+++ credential-verification-service/tests/integration_test_rest/verify.rs
@@ -1,6 +1,7 @@
 use crate::integration_test_helpers::{fixtures, server};
 use assert_matches::assert_matches;
 use concordium_rust_sdk::base::web3id::v1::CredentialVerificationMaterial;
+use concordium_rust_sdk::base::web3id::v1::anchor::PresentationVerifyFailure;
 use concordium_rust_sdk::common::cbor;
 use credential_verification_service::api_types::{VerificationResult, VerifyPresentationResponse};
 use reqwest::StatusCode;
@@ -10,7 +11,7 @@ use reqwest::StatusCode;
 async fn test_verify_account_based() {
     let handle = server::start_server();
     let global_context = fixtures::credentials::global_context();
-    let account_cred = fixtures::credentials::account_credentials_fixture(&global_context);
+    let account_cred = fixtures::credentials::account_credentials_fixture(&global_context, 1);
 
     let verify_fixture = fixtures::verify_request_account(&global_context, &account_cred);
 
@@ -29,13 +30,10 @@ async fn test_verify_account_based() {
 
     handle.node_client_stub().stub_account_credentials(
         account_cred.cred_id,
-        (
-            fixtures::chain::account_credentials(
-                &account_cred.cred_id,
-                verification_material.issuer,
-                verification_material.attribute_commitments.clone(),
-            ),
-            fixtures::chain::account_address(1),
+        fixtures::chain::account_credentials(
+            &account_cred.cred_id,
+            verification_material.issuer,
+            verification_material.attribute_commitments.clone(),
         ),
     );
 
@@ -157,8 +155,9 @@ async fn test_verify_fail() {
 
     assert_eq!(resp.status(), StatusCode::OK);
     let verify_response: VerifyPresentationResponse = resp.json().await.unwrap();
-    assert_matches!(verify_response.result, VerificationResult::Failed(err) => {
-        assert_eq!(err, "verification request anchor block hash not set in context");
+    assert_matches!(verify_response.result, VerificationResult::Failed(failure) => {
+        assert_eq!(failure.code, PresentationVerifyFailure::NoVraBlockHash);
+        assert_eq!(failure.message, "verification request anchor block hash not set in context");
     });
     assert!(verify_response.anchor_transaction_hash.is_none());
     assert_eq!(
@@ -172,5 +171,139 @@ async fn test_verify_fail() {
     assert_eq!(
         verify_response.verification_audit_record.presentation,
         verify_fixture.request.presentation
+    );
+}
+
+/// Test anchor not finalized
+#[tokio::test]
+async fn test_verify_anchor_not_finalized() {
+    let handle = server::start_server();
+    let global_context = fixtures::credentials::global_context();
+    let id_cred = fixtures::credentials::identity_credentials_fixture(&global_context);
+
+    let verify_fixture = fixtures::verify_request_identity(&global_context, &id_cred);
+
+    handle.node_client_stub().stub_block_item_status(
+        verify_fixture.anchor_txn_hash,
+        fixtures::chain::transaction_status_committed(
+            verify_fixture.anchor_txn_hash,
+            cbor::cbor_encode(&verify_fixture.anchor)
+                .unwrap()
+                .try_into()
+                .unwrap(),
+        ),
+    );
+
+    let resp = handle
+        .rest_client()
+        .post("verifiable-presentations/verify")
+        .json(&verify_fixture.request)
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::UNPROCESSABLE_ENTITY);
+    let resp_text = resp.text().await.unwrap();
+    assert!(
+        resp_text.contains("request anchor transaction") && resp_text.contains("not finalized"),
+        "response: {}",
+        resp_text
+    );
+}
+
+/// Test anchor not CBOR decodable
+#[tokio::test]
+async fn test_verify_anchor_not_decodable() {
+    let handle = server::start_server();
+    let global_context = fixtures::credentials::global_context();
+    let id_cred = fixtures::credentials::identity_credentials_fixture(&global_context);
+
+    let verify_fixture = fixtures::verify_request_identity(&global_context, &id_cred);
+
+    handle.node_client_stub().stub_block_item_status(
+        verify_fixture.anchor_txn_hash,
+        fixtures::chain::transaction_status_finalized(
+            verify_fixture.anchor_txn_hash,
+            vec![0].try_into().unwrap(),
+        ),
+    );
+
+    let resp = handle
+        .rest_client()
+        .post("verifiable-presentations/verify")
+        .json(&verify_fixture.request)
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::UNPROCESSABLE_ENTITY);
+    let resp_text = resp.text().await.unwrap();
+    assert!(
+        resp_text.contains("error decoding registered data"),
+        "response: {}",
+        resp_text
+    );
+}
+
+/// Test request anchor not found
+#[tokio::test]
+async fn test_verify_anchor_not_found() {
+    let handle = server::start_server();
+    let global_context = fixtures::credentials::global_context();
+    let id_cred = fixtures::credentials::identity_credentials_fixture(&global_context);
+
+    let verify_fixture = fixtures::verify_request_identity(&global_context, &id_cred);
+
+    let resp = handle
+        .rest_client()
+        .post("verifiable-presentations/verify")
+        .json(&verify_fixture.request)
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::UNPROCESSABLE_ENTITY);
+    let resp_text = resp.text().await.unwrap();
+    assert!(
+        resp_text.contains("request anchor transaction") && resp_text.contains("not found"),
+        "response: {}",
+        resp_text
+    );
+}
+
+/// Test account credential not found
+#[tokio::test]
+async fn test_verify_account_credential_not_found() {
+    let handle = server::start_server();
+    let global_context = fixtures::credentials::global_context();
+    let account_cred = fixtures::credentials::account_credentials_fixture(&global_context, 2);
+
+    let verify_fixture = fixtures::verify_request_account(&global_context, &account_cred);
+
+    handle.node_client_stub().stub_block_item_status(
+        verify_fixture.anchor_txn_hash,
+        fixtures::chain::transaction_status_finalized(
+            verify_fixture.anchor_txn_hash,
+            cbor::cbor_encode(&verify_fixture.anchor)
+                .unwrap()
+                .try_into()
+                .unwrap(),
+        ),
+    );
+
+    let resp = handle
+        .rest_client()
+        .post("verifiable-presentations/verify")
+        .json(&verify_fixture.request)
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(resp.status(), StatusCode::UNPROCESSABLE_ENTITY);
+    let resp_text = resp.text().await.unwrap();
+    assert!(
+        resp_text.contains("account credential") && resp_text.contains("not found"),
+        "response: {}",
+        resp_text
     );
 }